remove unused legacy mixer store code
