+++ mixer/pkg/config/store/BUILD
@@ -6,11 +6,9 @@ go_library(
     name = "go_default_library",
     srcs = [
         "convert.go",
-        "fsstore.go",
         "fsstore2.go",
         "memstore.go",
         "queue.go",
-        "store.go",
         "store2.go",
         "validator.go",
     ],
@@ -31,7 +29,6 @@ go_test(
         "fsstore2_test.go",
         "queue_test.go",
         "store2_test.go",
-        "store_test.go",
         "validator_test.go",
     ],
     library = ":go_default_library",
